+++ app/infrastructure/dto/marker_dto.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from uuid import UUID
 from pydantic import BaseModel
 
@@ -6,5 +7,5 @@ class MarkerDTO(BaseModel):
     marker_id: UUID
     incident_id: UUID
     direction: str
-    latitude: float
-    longitude: float 
+    latitude: Decimal
+    longitude: Decimal
